requirements.txt
newsapi-python>=0.2.7
pandas>=2.1.3
